refactoring query utilization
