+++ src/dune_analytics.py
@@ -12,24 +12,12 @@ from typing import Optional
 
 from requests import Session, Response
 
-from src.request import (
-    get_result_post,
-    find_result_post,
-    execute_query_post,
-    initiate_query_post,
-    Post,
-)
+from src.dune_query import DuneSQLQuery, Post
 from src.response import (
     validate_and_parse_dict_response,
     validate_and_parse_list_response,
 )
-from src.types import (
-    Network,
-    QueryParameter,
-    DuneRecord,
-    QueryResults,
-    DuneSQLQuery,
-)
+from src.types import DuneRecord, QueryResults
 
 log = logging.getLogger(__name__)
 logging.config.fileConfig(fname="logging.conf", disable_existing_loggers=True)
@@ -47,7 +35,6 @@ class DuneAnalytics:
         self,
         username: str,
         password: str,
-        query_id: int,
         max_retries: int = 2,
         ping_frequency: int = 5,
     ):
@@ -55,14 +42,12 @@ class DuneAnalytics:
         Initialize the object
         :param username: username for dune.xyz
         :param password: password for dune.xyz
-        :param query_id: existing integer query id owned `username`
         """
         self.csrf = None
         self.auth_refresh = None
         self.token = None
         self.username = username
         self.password = password
-        self.query_id = int(query_id)
         self.session = Session()
         self.max_retries = max_retries
         self.ping_frequency = ping_frequency
@@ -83,7 +68,6 @@ class DuneAnalytics:
         dune = DuneAnalytics(
             os.environ["DUNE_USER"],
             os.environ["DUNE_PASSWORD"],
-            int(os.environ["DUNE_QUERY_ID"]),
         )
         dune.login()
         dune.fetch_auth_token()
@@ -134,35 +118,35 @@ class DuneAnalytics:
         Initiates a new query.
         If no exception is raised, post was success!
         """
-        post_data = initiate_query_post(query)
+        post_data = query.initiate_query_post()
         response = self.post_dune_request(post_data)
         validate_and_parse_dict_response(response, post_data.key_map)
 
-    def execute_query(self) -> None:
+    def execute_query(self, query: DuneSQLQuery) -> None:
         """Executes query at query_id"""
-        post_data = execute_query_post(self.query_id)
+        post_data = query.execute_query_post()
         response = self.post_dune_request(post_data)
         validate_and_parse_dict_response(response, post_data.key_map)
 
-    def query_result_id(self) -> Optional[str]:
+    def query_result_id(self, query: DuneSQLQuery) -> Optional[str]:
         """
         Fetch the query result id for a query
         :return: string representation of integer result id
         """
-        post_data = get_result_post(self.query_id)
+        post_data = query.get_result_post()
         response = self.post_dune_request(post_data)
         response_data = validate_and_parse_dict_response(response, post_data.key_map)
 
         return response_data["get_result"].get("result_id", None)
 
-    def get_results(self) -> list[DuneRecord]:
+    def get_results(self, query: DuneSQLQuery) -> list[DuneRecord]:
         """Fetch the result for a query by id"""
-        result_id = self.query_result_id()
+        result_id = self.query_result_id(query)
         while not result_id:
             time.sleep(self.ping_frequency)
             log.debug("Awaiting results... ")
-            result_id = self.query_result_id()
-        post_data = find_result_post(result_id)
+            result_id = self.query_result_id(query)
+        post_data = DuneSQLQuery.find_result_post(result_id)
         response = self.post_dune_request(post_data)
         response_data = validate_and_parse_list_response(response, post_data.key_map)
         return QueryResults(response_data).data
@@ -180,44 +164,26 @@ class DuneAnalytics:
 
         return response
 
-    def execute_and_await_results(self) -> list[DuneRecord]:
+    def execute_and_await_results(self, query: DuneSQLQuery) -> list[DuneRecord]:
         """
         Executes query by ID and awaits completion.
         :return: parsed list of dict records returned from query
         """
-        self.execute_query()
-        data_set = self.get_results()
+        self.execute_query(query)
+        data_set = self.get_results(query)
         log.info(f"got {len(data_set)} records from last query")
         return data_set
 
-    def fetch(
-        self,
-        query_str: str,
-        network: Network,
-        name: str,
-        parameters: Optional[list[QueryParameter]] = None,
-    ) -> list[DuneRecord]:
+    def fetch(self, query: DuneSQLQuery) -> list[DuneRecord]:
         """
-        Pushes new query and executes, awaiting query completion
-        :param query_str: sql string to execute
-        :param network: Network enum variant
-        :param name: optional name of what is being fetched (for logging)
-        :param parameters: optional parameters to be included in query
-        :return: list of records as dictionaries
+        Pushes new query, executes and awaiting query completion
+        :return: list query records as dictionaries
         """
-        log.info(f"Fetching {name} on {network}...")
-        self.initiate_query(
-            DuneSQLQuery(
-                query_id=self.query_id,
-                raw_sql=query_str,
-                network=network,
-                name="Auto Generated Query",
-                parameters=parameters or [],
-            )
-        )
+        log.info(f"Fetching {query.name} on {query.network}...")
+        self.initiate_query(query)
         for _ in range(0, self.max_retries):
             try:
-                return self.execute_and_await_results()
+                return self.execute_and_await_results(query)
             except RuntimeError as err:
                 log.warning(
                     f"failed with {err}. Re-establishing connection and trying again"
