+++ src/example/sample_fetch.py
@@ -1,9 +1,12 @@
 """Sample Fetch script from DuneAnalytics"""
 from __future__ import annotations
+
+import os
 from dataclasses import dataclass
 from datetime import datetime
 
 from src.dune_analytics import DuneAnalytics
+from src.dune_query import DuneSQLQuery
 from src.types import Network, QueryParameter
 from src.util import open_query
 
@@ -31,8 +34,8 @@ class Record:
 
 def fetch_records(dune: DuneAnalytics) -> list[Record]:
     """Initiates and executes Dune query, returning results as Python Objects"""
-    results = dune.fetch(
-        query_str=open_query("./src/example/sample_query.sql"),
+    sample_query = DuneSQLQuery(
+        raw_sql=open_query("./src/example/sample_query.sql"),
         name="Sample Query",
         network=Network.MAINNET,
         parameters=[
@@ -40,7 +43,9 @@ def fetch_records(dune: DuneAnalytics) -> list[Record]:
             QueryParameter.date_type("DateParam", datetime(2022, 3, 10, 12, 30, 30)),
             QueryParameter.text_type("TextParam", "aba"),
         ],
+        query_id=int(os.environ["DUNE_QUERY_ID"]),
     )
+    results = dune.fetch(sample_query)
     return [Record.from_dict(row) for row in results]
