+++ tests/e2e/test_dune_api.py
@@ -1,7 +1,9 @@
+import os
 import unittest
 
-from src.dune_analytics import DuneAnalytics, Network, QueryParameter
-from src.util import open_query
+from src.dune_analytics import DuneAnalytics
+from src.dune_query import DuneSQLQuery
+from src.types import Network, QueryParameter
 
 
 class TestDuneAnalytics(unittest.TestCase):
@@ -12,22 +14,24 @@ class TestDuneAnalytics(unittest.TestCase):
         - initiate_new_query
         - execute_and_await_results
         - execute_query
-        - handle_dune_request
+        - post_dune_request
         essentially all the methods of the API
         """
         dune = DuneAnalytics.new_from_environment()
         five, one = 5, 1
         parameter_name = "IntParameter"
         column_name = "value"
-        res = dune.fetch(
+        query = DuneSQLQuery(
             # Note that consecutive double brace brackets in formatted strings
             # become single brace brackets, so this query is
             # select 5 - '{{IntParameter}}' as value
-            query_str=f"select {five} - '{{{{{parameter_name}}}}}' as {column_name}",
+            raw_sql=f"select {five} - '{{{{{parameter_name}}}}}' as {column_name}",
             network=Network.MAINNET,
             parameters=[QueryParameter.number_type(parameter_name, one)],
             name="Test Fetch",
+            query_id=int(os.environ["DUNE_QUERY_ID"]),
         )
+        res = dune.fetch(query)
         self.assertEqual(len(res), 1)
         self.assertEqual(res[0][column_name], five - one)
